Split tests/dataset/test_dataset.py into seven modules
4284 lines and 426 of the suite's ~850 tests lived in one file, so there
was no smaller unit to run while iterating, and pytest-split could not
balance around it.

This is pure movement: all 153 top-level definitions are byte-identical
before and after, verified by comparing their parsed source. Shared
helpers and constants move to `_dataset_helpers.py`, the three
downsampling tests join the existing `test_downsampling.py`, and the
rest are grouped by what they exercise:

  test_dataset_create.py        creating, opening and default mags
  test_dataset_properties.py    OME-NGFF metadata, view config, segment ids
  test_dataset_io.py            reads, writes, axes and bounds
  test_dataset_layers.py        get_or_add, rename, delete, add_layer_like
  test_dataset_copy.py          refs, symlinks, copies and conversions
  test_dataset_chunking.py      chunking, compression, rechunking, zarr3
  test_dataset_remote_paths.py  remote and symlinked paths

An identical pass/fail set before and after is the whole verification.
